Added handling of custom "sidefx::noActivateOnButtonPress" property.
When the property is set on a window then clicking on the window
does not activate/raise it (Linux only).  This leaves it up to the
window manager to decide if/when the window should be raised and gain
focus.

There is an unresolved bug, QTBUG-40202, which touches on the issue
of automatically activating a window when it is clicked on.  In the bug
there was a proposed change of removing the activating code altogether
but that change was not accepted.

This change takes a less intrusive approach by disabling the
activating code only for windows that request not to be activated.