Revert "widgets: Mark widgets as needing flush during painting"
This reverts commit 1e310b5a4508f25a59f0e54aea8636f38f17db44.

The reverted commit caused 'stuttering' when resizing windows on KDE.
The problem only occurred in Houdini windows possibly because the
windows are OpenGL-based and render straight to the draw buffer instead
of to FBOs.  I ran some tests and found that the problem did not occur
in regular QWidget windows and QOpenGLWidget windows.