Do not adjust popup menus contained in graphics views.
This fixes a bug where a popup menu is positioned outside of the
graphics view area when opened for the very first time.  On subsequent
opens the menu is positioned correctly because after the first open,
QGraphicsProxyWidget sets the Qt::WA_DontShowOnScreen attribute which
also disables desktop adjustments.

The behavior of setting the Qt::WA_DontShowOnScreen attribute *after*
the menu is positioned for the very first time is the same in Qt4.
What's different in Qt5 and why it broke in Qt5 is that the
QWidget::mapToGlobal() method has changed.  In Qt4, mapToGlobal() mapped
to the 'global' position relative to the containing graphics view while
Qt5 maps to the global position of the desktop.  This means that in Qt5
the requested menu position is in global space but the adjusted desktop
is in graphics view space causing bad geometry comparisons and
repositions.