Committing our internal patch to modify Qt to bracket QKeyEvents with custom pre/post key events for passing extended information about what a key event would look like if modifiers were ignored.
The generation of these custom events is enabled through a custom platform
function queried by the token "sidefxSetBracketKeyEventTypes".  The custom
function takes two integer arguments, the custom event types that have been
registered via QEvent::registerEventType() for these two event types.

Note that the H19.5 version of this patch contains two bugs that were fixed
for the H20.0 version.

1. QXcbKeyboard::updateKeymap() did not correctly copy the numlock modifier
   state when bootstrapping m_sidefx_xkbStateNoModifiers.
2. The OSX 10.15+ codepath incorrectly assumed that encoding of non-symbol
   keys by NSEvent::characters(byApplyingModifiers:) was done the same way
   as that by UCKeyTranslate().

H19.5 does not make use of the data passed through these custom events, so
these bugs don't affect it, but H20.0 and up should not use this version of
the patch.