Rollback QTBUG-85431 fix, but only for macOS.
The fix prevents MouseMove events from reaching Qt SubWindows
in some cases.  On Mac, MouseMove events are sometimes generated
for both the parent Window and child SubWindow for the same move.
The parent Window sees the move first but because of the fix,
the child SubWindow never sees the move.  This caused focus issues
in Houdini where the viewport (subwindow) was unaware it had
mouse focus.

Note that this change potentially reintroduces the QTBUG-85431
on Mac.

This is a manual cherry-pick of
3c1b71712236a155022f4269b54d88536d07c5da.