Fix header view styling such that centered text is actually centered.
308845cd720969a5f3456caa8cf72023f120aecf broke this behavior
when it took into account the space of the sorter arrow in the
text size calculation.  The commit was meant to fix overlapping
issues for right-aligned and bottom-aligned text,
according to the commit log message, so we limit the scope of the
change so that it only applies when the text alignment is set to
either AlignRight or AlignBottom, depending on the header orientation.